back/requirements.txt
psycopg2-binary
redis
python-dotenv
fastapi
uvicorn 